+++ prisma/migrations/20260913161932_add_subscription_activation_email/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Subscription" ADD COLUMN     "activationEmailSentAt" TIMESTAMP(3);